backend/requirements.txt
tradingeconomics==0.2.3
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pytest==7.4.4
httpx==0.26.0
pandas==2.2.0 